+++ R/summarizor.R
@@ -13,6 +13,7 @@
 #' @param overall_label label to use as overall label
 #' @seealso [fmt_summarizor()], [labelizor()]
 #' @examples
+#' \dontrun{
 #' z <- summarizor(CO2[-c(1, 4)],
 #'   by = "Treatment",
 #'   overall_label = "Overall"
@@ -73,6 +74,7 @@
 #'
 #' ft_2 <- as_flextable(tab_2, separate_with = "variable")
 #' ft_2
+#' }
 #' @export
 summarizor <- function(
     x, by = character(),
@@ -142,6 +144,7 @@ summarizor <- function(
 #' @param ... arguments for [as_flextable.tabulator()]
 #' @family as_flextable methods
 #' @examples
+#' \dontrun{
 #' z <- summarizor(CO2[-c(1, 4)],
 #'   by = "Treatment",
 #'   overall_label = "Overall"
@@ -152,6 +155,7 @@ summarizor <- function(
 #'   as_chunk("\t"))
 #' ft_1 <- autofit(ft_1)
 #' ft_1
+#' }
 as_flextable.summarizor <- function(x, ...) {
 
   tab <- tabulator(
@@ -351,6 +355,8 @@ fmt_n_percent <- function(n, pct, digit = 1){
 #' @description The function formats counts as `\n(N=XX)`. This helper
 #' function is used to add counts in columns titles.
 #' @param n count values
+#' @param newline indicates to prefix the text with a new line
+#' (sof return).
 #' @seealso [tabulator()], [mk_par()]
 #' @family text formatter functions
 #' @examples
@@ -364,9 +370,15 @@ fmt_n_percent <- function(n, pct, digit = 1){
 #'   value = as_chunk(fmt_header_n(200)))
 #' ft_1 <- autofit(ft_1)
 #' ft_1
-fmt_header_n <- function(n){
+fmt_header_n <- function(n, newline = TRUE){
   z1 <- character(length(n))
-  z1[!is.na(n)] <- sprintf("\n(N=%s)", fmt_int(n[!is.na(n)]))
+
+  mask <- "\n(N=%s)"
+  if (!newline) {
+    mask <- "(N=%s)"
+  }
+
+  z1[!is.na(n)] <- sprintf(mask, fmt_int(n[!is.na(n)]))
   z1
 }